Trafos: Fix small bugs, rename dir
